Update address.h
long -> uint32_t

+++ omap/kernel/arch/address.h
@@ -6,10 +6,10 @@
 #define address(X, Y) __CONCRETE_ADDRESS(X, Y)
 
 //address of a register
-typedef volatile long* address;
+typedef volatile uint32_t* address;
 
 //base address
-typedef volatile long base_address;
+typedef volatile uint32_t base_address;
 
 //offset to base address
 typedef volatile uint16_t offset;
